requirements.txt
flask>=3.0.0
requests>=2.31.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-flask>=0.43b0
opentelemetry-instrumentation-requests>=0.43b0
opentelemetry-exporter-otlp-proto-http>=1.22.0
opentelemetry-semantic-conventions>=0.43b0
prometheus-client>=0.19.0
